requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
trafilatura==2.0.0
readability-lxml==0.8.4.1
pyyaml==6.0.2
mysql-connector-python==9.1.0
vaderSentiment==3.3.2
googlenewsdecoder==0.1.7
